Refactor home page bubble
