+++ src/components/home/bubble.module.css
@@ -1,0 +1,30 @@
+.container {
+  position: absolute;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 13vw;
+  aspect-ratio: 1/1;
+  border-radius: 50%;
+  background-image: url('/home/bubble_bg.png');
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center;
+  cursor: pointer;
+}
+
+.image {
+  width: 65%;
+  height: 65%;
+}
+
+.fade_mask {
+  position: absolute;
+  inset: 0;
+  background: linear-gradient(
+    to bottom,
+    rgba(255, 255, 255, 0) 0%,
+    rgba(255, 255, 255, 1) 100%
+  );
+  pointer-events: none;
+}
